re-design shell 2
