+++ src/patterns/shells/shell1.js
@@ -2,88 +2,85 @@
  * WordPress dependencies.
  */
 import {
-    Icon,
     Card,
     CardHeader,
     CardBody,
-    TabPanel,
     Button,
     __experimentalHStack as HStack,
-    __experimentalText as Text,
+    __experimentalVStack as VStack,
     __experimentalHeading as Heading,
 } from "@wordpress/components";
-import { home, cog, plugins, key, external,trendingUp } from "@wordpress/icons";
+import { home, cog, plugins, key, external, trendingUp, menu, close } from "@wordpress/icons";
+import { useViewportMatch } from '@wordpress/compose';
 
 /**
- * Render Shell 1
+ * Internal dependencies.
  */
+import React, { useState } from 'react';
+
+// Consolidated button data
+const buttonData = [
+    { icon: home, label: 'Dashboard' },
+    { icon: cog, label: 'Settings' },
+    { icon: plugins, label: 'Addons' },
+    { icon: key, label: 'License' },
+    { icon: external, label: 'Support' },
+    { icon: trendingUp, label: 'Gro Pro', variant: 'primary' }
+];
+
 function Shell1() {
+    const isMobile = !useViewportMatch('mobile');
+    const [showButtons, setShowButtons] = useState(false);
+    const [display, setDisplay] = useState('flex');
+    const [icon, setIcon] = useState(menu);
+
+    const handleButtonClick = () => {
+        setIcon(prevIcon => (prevIcon === menu ? close : menu));
+        setShowButtons(prevState => !prevState);
+        setDisplay('none')
+    };
+
     return (
         <Card>
             <CardHeader>
                 <HStack>
                     <Heading size={20}>WPUI</Heading>
-                    <TabPanel
-                        tabs={[
-                            {
-                                name: 'tab1',
-                                title: (
-                                    <HStack spacing={2}>
-                                        <Icon icon={home} />
-                                        <Text>Dashboard</Text>
-                                    </HStack>
-                                ),
-                            },
-                            {
-                                name: 'tab2',
-                                title: (
-                                    <HStack spacing={2}>
-                                        <Icon icon={cog} />
-                                        <Text>Settings</Text>
-                                    </HStack>
-                                ),
-                            },
-                            {
-                                name: 'tab3',
-                                title: (
-                                    <HStack spacing={2}>
-                                        <Icon icon={plugins} />
-                                        <Text> Addons</Text>
-                                    </HStack>
-                                ),
-                            },
-                            {
-                                name: 'tab4',
-                                title: (
-                                    <HStack spacing={2}>
-                                        <Icon icon={key} />
-                                        <Text>License</Text>
-                                    </HStack>
-                                ),
-                            },
-                            {
-                                name: 'tab5',
-                                title: (
-                                    <HStack spacing={2}>
-                                        <Icon icon={external} />
-                                        <Text>Support</Text>
-                                    </HStack>
-                                ),
-                            },
-                        ]}
-                    >
-                        {(tab) => <div></div>}
-                    </TabPanel>
-                    <Button variant="primary" icon={trendingUp}>Gro Pro</Button>
+                    {isMobile ? (
+                        <Button icon={icon} onClick={handleButtonClick}></Button>
+                    ) : (
+                        <>
+                            <HStack expanded={false} spacing={5}>
+                                {buttonData.slice(0, -1).map((btn, index) => (
+                                    <Button key={index} icon={btn.icon}>
+                                        {btn.label}
+                                    </Button>
+                                ))}
+                            </HStack>
+                            <Button variant="primary" icon={trendingUp}>Gro Pro</Button>
+                        </>
+                    )}
                 </HStack>
             </CardHeader>
-            <CardBody style={{ padding: "100px" }}>
-                {/*Display Your Content Here*/}
+            {isMobile && showButtons && (
+                <VStack style={{ padding: '12px' }} expanded={false} spacing={5}>
+                    {buttonData.map((btn, index) => (
+                        <Button
+                            key={index}
+                            icon={btn.icon}
+                            variant={btn.variant}
+                            style={btn.label === 'Gro Pro' ? { display: 'flex', justifyContent: 'center' } : {}}
+                        >
+                            {btn.label}
+                        </Button>
+                    ))}
+                </VStack>
+            )}
+            <CardBody style={{ padding: "100px", display: display, justifyContent: 'center' }}>
+                {/*Display Your Content Here*/}This is Output Panel
             </CardBody>
         </Card>
     );
 }
-
 
 // @meta-start
 Shell1.meta = {
